changed the directory structure a tad for this new feature
all movement nodes now have an ID associated with them to be referenced
in the network. This ID is generated from the position of the nodes script
on the disc. This way we do not have to manually input the ids YET.
Before game release these ID's should be scripted and baked into the code

for development this solution works fine